+++ src/mplchart/primitives/areaplot.py
@@ -34,6 +34,7 @@ class AreaPlot(BindingPrimitive):
         alpha: float | None = None,
         label: str | None = None,
         legend: bool = True,
+        pane: str | None = None,
     ):
 
         super().__init__(indicator)
@@ -41,9 +42,10 @@ class AreaPlot(BindingPrimitive):
         self.alpha = alpha
         self.label = label
         self.legend = legend
+        self.pane = pane
 
     def apply_to_chart(self, chart):
-        ax = chart.canvas.get_axes()
+        ax = chart.canvas.get_axes(self.pane)
 
         result = chart.view.eval(self.required_indicator())